fix(test): #2462 — shared canvas_pixel_probe.hpp; repair shadow test compile
test_canvas_widget_shadow.cpp (extracted from test_canvas_widget.cpp in
the Phase 5 P5-3 split, #2418) referenced sample_pixel / Pixel without
a definition — it failed to compile under a Skia-enabled build. The
P5-3 split validated the canvas tests without Skia, so the Skia-gated
branch of the extracted file never compiled. This is the loose end
pulp #2462 tracked.

Fix per #2462's suggested approach: extract the raster-pixel probe
(Pixel struct + sample_pixel) into a shared test/canvas_pixel_probe.hpp
included by all three canvas test files, instead of a third duplicate
copy. test_canvas_widget.cpp and test_canvas2d_shim.cpp drop their
identical local copies; test_canvas_widget_shadow.cpp gains the
declaration it was missing. The header is PULP_HAS_SKIA-gated.

Verified WITH Skia (-DSKIA_DIR, PULP_HAS_SKIA=1) — the gap that hid
this originally:
  * full `cmake --build -- -k` keep-going build: 461 targets, exit 0,
    zero errors — no other broken test targets behind this one
  * pulp-test-canvas-widget-shadow → 30 assertions / 4 cases (the 2
    Skia-gated shadow tests now compile + run, previously skipped)
  * pulp-test-canvas-widget (190/44), -sanitize (28/5),
    test-canvas2d-shim (167/53) all green

Note: #2462's systemic item (PR CI does not `make all`, so test
targets rot undetected) is a separate nightly-full-build task and
stays open — this PR fixes only the file.
